seperated CRUD logic into different classes, removed all update logic
